@@ -93,26 +101,34 @@
     "concurrently": "^7.0.0",
     "glob": "^10.2.2",
     "husky": "^8.0.0",
+    "jest": "^30.0.0",
+    "jest-circus": "^30.0.0",
+    "jest-environment-node": "^30.0.0",
     "jest-image-snapshot": "^6.2.0",
+    "jest-runner": "^30.0.0",
     "lint-staged": "^13.0.3",
-    "node-fetch": "^2",
+    "pathe": "^2.0.3",
     "pkg-up": "^5.0.0",
+    "playwright": ">=1.24.0",
+    "playwright-chromium": ">=1.24.0",
     "prettier": "^2.8.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "read-pkg-up": "^7.0.1",
-    "storybook": "next",
+    "storybook": "^10.0.0-beta.10",
     "tempy": "^1.0.1",
     "ts-dedent": "^2.0.0",
-    "ts-jest": "^29.0.0",
-    "tsup": "^6.5.0",
-    "typescript": "~4.9.4",
-    "vite": "^6.3.2",
+    "ts-jest": "^29.4.0",
+    "tsup": "^8.5.0",
+    "typescript": "^5.8.3",
+    "vite": "^7.0.5",
+    "vitest": "^3.2.4",
     "wait-on": "^7.2.0"
   },
   "peerDependencies": {
-    "storybook": "^0.0.0-0 || ^8.2.0 || ^9.0.0 || ^9.1.0-0 || ^9.2.0-0"
+    "storybook": "^10.0.0 || ^10.0.0-0"
   },
+  "packageManager": "yarn@4.5.1",
   "engines": {
     "node": ">=20.0.0"
   },
