@@ -1,17 +1,33 @@
-const { transformSync: swcTransform } = require('@swc/core');
-const { transformPlaywright } = require('../dist');
+import { transform as swcTransform } from '@swc/core';
+import { transformPlaywright } from '../dist/index.js';
 
-module.exports = {
-  process(src, filename) {
-    const csfTest = transformPlaywright(src, filename);
-
-    const result = swcTransform(csfTest, {
+// Only export async version - force Jest to use it
+async function processAsync(src, filename) {
+  try {
+    const csfTest = await transformPlaywright(src, filename);
+    // This swc transform might not be needed
+    const result = await swcTransform(csfTest, {
       filename,
+      isModule: true,
       module: {
-        type: 'commonjs',
+        type: 'es6',
+      },
+      jsc: {
+        parser: {
+          syntax: 'typescript',
+          tsx: true,
+        },
+        target: 'es2015',
       },
     });
 
     return { code: result ? result.code : src };
-  },
+  } catch (error) {
+    console.error('Transform error:', error);
+    return { code: src };
+  }
+}
+
+export default {
+  processAsync,
 };
