@@ -1,27 +1,31 @@
 // !!! This file is used as an override to the test-runner configuration for this repo only !!!
 // If you want to create your own override for your project, run test-storybook eject instead
-const path = require('path');
+import path from 'path';
+import { fileURLToPath } from 'url';
+
 // we override the path here so that when running the test-runner locally, it resolves to local files instead when calling require.resolve
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 process.env.STORYBOOK_TEST_RUNNER_PATH = path.resolve(__dirname);
 
-const { getJestConfig } = require('./dist');
+import { getJestConfig } from './dist/index.js';
 
 const testRunnerConfig = getJestConfig();
 
 /**
  * @type {import('@jest/types').Config.InitialOptions}
  */
-module.exports = {
+export default {
   ...testRunnerConfig,
   cacheDirectory: 'node_modules/.cache/storybook/test-runner',
   transform: {
-    '^.+\\.(story|stories)\\.[jt]sx?$': './playwright/transform',
+    '^.+\\.(story|stories)\\.[jt]sx?$': './playwright/transform.js',
     '^.+\\.[jt]sx?$': '@swc/jest',
   },
-  globalSetup: './playwright/global-setup.js',
-  globalTeardown: './playwright/global-teardown.js',
+  globalSetup: './dist/jest-playwright-entries/setup.js',
+  globalTeardown: './dist/jest-playwright-entries/teardown.js',
   testEnvironment: './playwright/custom-environment.js',
   setupFilesAfterEnv: ['./playwright/jest-setup.js'],
+  // transformIgnorePatterns for the storybook package and make sure swc jest ignores it too
   // use local build when the package is referred
   moduleNameMapper: {
     '@storybook/test-runner': '<rootDir>/dist/index.js',
