@@ -1,9 +1,10 @@
 import { TestPrefixer, TransformOptions, transformCsf } from './transformCsf';
 import { testPrefixer } from '../playwright/transformPlaywright';
 import template from '@babel/template';
+import { describe, it, expect } from 'vitest';
 
-describe('transformCsf', () => {
-  it('inserts a no-op test if there are no stories', () => {
+describe('transformCsf', async () => {
+  it('inserts a no-op test if there are no stories', async () => {
     const csfCode = `
       export default {
         title: 'Button',
@@ -11,65 +12,24 @@ describe('transformCsf', () => {
     `;
     const expectedCode = `describe.skip('Button', () => { it('no-op', () => {}) });`;
 
-    const result = transformCsf(csfCode, { insertTestIfEmpty: true } as TransformOptions);
+    const result = await transformCsf(csfCode, { insertTestIfEmpty: true } as TransformOptions);
 
     expect(result).toEqual(expectedCode);
   });
 
-  it('returns empty result if there are no stories', () => {
+  it('returns empty result if there are no stories', async () => {
     const csfCode = `
       export default {
         title: 'Button',
       };
     `;
 
-    const result = transformCsf(csfCode, { testPrefixer });
+    const result = await transformCsf(csfCode, { testPrefixer });
 
     expect(result).toMatchSnapshot();
   });
 
-  it('calls the testPrefixer function for each test', () => {
-    const csfCode = `
-      export default {
-        title: 'Button',
-        parameters: {
-          play: {
-            steps: [
-              { id: 'step1', action: 'click', target: 'button' },
-              { id: 'step2', action: 'click', target: 'button' },
-            ],
-          },
-        },
-      };
-      export const Primary = () => '<button>Primary</button>';
-    `;
-
-    const result = transformCsf(csfCode, { testPrefixer });
-
-    expect(result).toMatchSnapshot();
-  });
-
-  it('calls the beforeEachPrefixer function once', () => {
-    const csfCode = `
-      export default {
-        title: 'Button',
-        parameters: {
-          play: {
-            steps: [
-              { id: 'step1', action: 'click', target: 'button' },
-              { id: 'step2', action: 'click', target: 'button' },
-            ],
-          },
-        },
-      };
-      export const Primary = () => '<button>Primary</button>';
-    `;
-    const result = transformCsf(csfCode, { testPrefixer, beforeEachPrefixer: undefined });
-
-    expect(result).toMatchSnapshot();
-  });
-
-  it('clears the body if clearBody option is true', () => {
+  it('calls the testPrefixer function for each test', async () => {
     const csfCode = `
       export default {
         title: 'Button',
@@ -85,12 +45,53 @@ describe('transformCsf', () => {
       export const Primary = () => '<button>Primary</button>';
     `;
 
-    const result = transformCsf(csfCode, { testPrefixer, clearBody: true });
+    const result = await transformCsf(csfCode, { testPrefixer });
 
     expect(result).toMatchSnapshot();
   });
 
-  it('executes beforeEach code before each test', () => {
+  it('calls the beforeEachPrefixer function once', async () => {
+    const csfCode = `
+      export default {
+        title: 'Button',
+        parameters: {
+          play: {
+            steps: [
+              { id: 'step1', action: 'click', target: 'button' },
+              { id: 'step2', action: 'click', target: 'button' },
+            ],
+          },
+        },
+      };
+      export const Primary = () => '<button>Primary</button>';
+    `;
+    const result = await transformCsf(csfCode, { testPrefixer, beforeEachPrefixer: undefined });
+
+    expect(result).toMatchSnapshot();
+  });
+
+  it('clears the body if clearBody option is true', async () => {
+    const csfCode = `
+      export default {
+        title: 'Button',
+        parameters: {
+          play: {
+            steps: [
+              { id: 'step1', action: 'click', target: 'button' },
+              { id: 'step2', action: 'click', target: 'button' },
+            ],
+          },
+        },
+      };
+      export const Primary = () => '<button>Primary</button>';
+    `;
+
+    const result = await transformCsf(csfCode, { testPrefixer, clearBody: true });
+
+    expect(result).toMatchSnapshot();
+  });
+
+  it('executes beforeEach code before each test', async () => {
     const code = `
     export default {
         title: 'Button',
@@ -114,7 +115,10 @@ describe('transformCsf', () => {
       console.log({ id: %%id%%, title: %%title%%, name: %%name%%, storyExport: %%storyExport%% });
       async () => {}`) as unknown as TestPrefixer;
 
-    const result = transformCsf(code, { beforeEachPrefixer, testPrefixer } as TransformOptions);
+    const result = await transformCsf(code, {
+      beforeEachPrefixer,
+      testPrefixer,
+    } as TransformOptions);
 
     expect(result).toMatchSnapshot();
   });
