@@ -1,13 +1,19 @@
 import dedent from 'ts-dedent';
 import path from 'path';
 import * as storybookMain from '../util/getStorybookMain';
+import { describe, beforeEach, expect, it, vi } from 'vitest';
 
 import { transformPlaywright } from './transformPlaywright';
 
-jest.mock('storybook/internal/common', () => ({
-  ...jest.requireActual('storybook/internal/common'),
-  getProjectRoot: jest.fn(() => '/foo/bar'),
-  normalizeStories: jest.fn(() => [
+vi.mock('storybook/internal/preview-api', async (importOriginal) => ({
+  ...(await importOriginal()),
+  userOrAutoTitle: vi.fn(() => 'Example/Header'),
+}));
+
+vi.mock('storybook/internal/common', async (importOriginal) => ({
+  ...(await importOriginal()),
+  getProjectRoot: vi.fn(() => '/stories/basic'),
+  normalizeStories: vi.fn(() => [
     {
       titlePrefix: 'Example',
       files: '**/*.stories.@(mdx|tsx|ts|jsx|js)',
@@ -18,7 +24,7 @@ jest.mock('storybook/internal/common', () => ({
   ]),
 }));
 
-jest.mock('../util/getTestRunnerConfig');
+vi.mock('../util/getTestRunnerConfig');
 
 expect.addSnapshotSerializer({
   print: (val: unknown) => (typeof val === 'string' ? val.trim() : String(val)),
@@ -28,16 +34,18 @@ expect.addSnapshotSerializer({
 describe('Playwright', () => {
   const filename = './stories/basic/Header.stories.js';
   beforeEach(() => {
-    const relativeSpy = jest.spyOn(path, 'relative');
+    const relativeSpy = vi.spyOn(path, 'relative');
     relativeSpy.mockReturnValueOnce('stories/basic/Header.stories.js');
-    jest.spyOn(storybookMain, 'getStorybookMain').mockImplementation(() => ({
-      stories: [
-        {
-          directory: '../stories/basic',
-          titlePrefix: 'Example',
-        },
-      ],
-    }));
+    vi.spyOn(storybookMain, 'getStorybookMain').mockImplementation(() =>
+      Promise.resolve({
+        stories: [
+          {
+            directory: '../stories/basic',
+            titlePrefix: 'Example',
+          },
+        ],
+      })
+    );
 
     delete process.env.STORYBOOK_INCLUDE_TAGS;
     delete process.env.STORYBOOK_EXCLUDE_TAGS;
@@ -46,8 +54,8 @@ describe('Playwright', () => {
   });
 
   describe('tag filtering mechanism', () => {
-    it('should include all stories when there is no tag filtering', () => {
-      expect(
+    it('should include all stories when there is no tag filtering', async () => {
+      await expect(
         transformPlaywright(
           dedent`
         export default { title: 'foo/bar', component: Button };
@@ -56,136 +64,134 @@ describe('Playwright', () => {
       `,
           filename
         )
-      ).toMatchInlineSnapshot(`
-        if (!require.main) {
-          describe("Example/foo/bar", () => {
-            describe("A", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--a",
-                    title: "Example/foo/bar",
-                    name: "A"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--a"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
+      ).resolves.toMatchInlineSnapshot(`
+        describe("Example/Header", () => {
+          describe("A", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--a",
+                  title: "Example/Header",
+                  name: "A"
                 };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
                 try {
-                  await testFn();
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--a"
+                  });
                 } catch (err) {
                   if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"A"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
+                    throw err;
                   } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
                     throw err;
                   }
                 }
-              });
-            });
-            describe("B", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--b",
-                    title: "Example/foo/bar",
-                    name: "B"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--b"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
-                  await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"B"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
                 }
-              });
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"A"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
             });
           });
-        }
+          describe("B", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--b",
+                  title: "Example/Header",
+                  name: "B"
+                };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
+                try {
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--b"
+                  });
+                } catch (err) {
+                  if (err.toString().includes('Execution context was destroyed')) {
+                    throw err;
+                  } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
+                    throw err;
+                  }
+                }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"B"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
+            });
+          });
+        });
       `);
     });
-    it('should exclude stories when excludeTags matches', () => {
+    it('should exclude stories when excludeTags matches', async () => {
       process.env.STORYBOOK_EXCLUDE_TAGS = 'exclude-test';
-      expect(
+      await expect(
         transformPlaywright(
           dedent`
         export default { title: 'foo/bar', component: Button };
@@ -194,76 +200,74 @@ describe('Playwright', () => {
       `,
           filename
         )
-      ).toMatchInlineSnapshot(`
-        if (!require.main) {
-          describe("Example/foo/bar", () => {
-            describe("B", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--b",
-                    title: "Example/foo/bar",
-                    name: "B"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--b"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
+      ).resolves.toMatchInlineSnapshot(`
+        describe("Example/Header", () => {
+          describe("B", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--b",
+                  title: "Example/Header",
+                  name: "B"
                 };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
                 try {
-                  await testFn();
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--b"
+                  });
                 } catch (err) {
                   if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"B"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
+                    throw err;
                   } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
                     throw err;
                   }
                 }
-              });
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"B"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
             });
           });
-        }
+        });
       `);
     });
-    it('should skip stories when skipTags matches', () => {
+    it('should skip stories when skipTags matches', async () => {
       process.env.STORYBOOK_SKIP_TAGS = 'skip-test';
-      expect(
+      await expect(
         transformPlaywright(
           dedent`
         export default { title: 'foo/bar', component: Button };
@@ -272,134 +276,132 @@ describe('Playwright', () => {
       `,
           filename
         )
-      ).toMatchInlineSnapshot(`
-        if (!require.main) {
-          describe("Example/foo/bar", () => {
-            describe("A", () => {
-              it.skip("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--a",
-                    title: "Example/foo/bar",
-                    name: "A"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--a"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
+      ).resolves.toMatchInlineSnapshot(`
+        describe("Example/Header", () => {
+          describe("A", () => {
+            it.skip("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--a",
+                  title: "Example/Header",
+                  name: "A"
                 };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
                 try {
-                  await testFn();
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--a"
+                  });
                 } catch (err) {
                   if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"A"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
+                    throw err;
                   } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
                     throw err;
                   }
                 }
-              });
-            });
-            describe("B", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--b",
-                    title: "Example/foo/bar",
-                    name: "B"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--b"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
-                  await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"B"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
                 }
-              });
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"A"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
             });
           });
-        }
+          describe("B", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--b",
+                  title: "Example/Header",
+                  name: "B"
+                };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
+                try {
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--b"
+                  });
+                } catch (err) {
+                  if (err.toString().includes('Execution context was destroyed')) {
+                    throw err;
+                  } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
+                    throw err;
+                  }
+                }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"B"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
+            });
+          });
+        });
       `);
     });
-    it('should work in conjunction with includeTags, excludeTags and skipTags', () => {
+    it('should work in conjunction with includeTags, excludeTags and skipTags', async () => {
       process.env.STORYBOOK_INCLUDE_TAGS = 'play,design,global-tag';
       process.env.STORYBOOK_SKIP_TAGS = 'skip';
       process.env.STORYBOOK_EXCLUDE_TAGS = 'exclude';
@@ -411,7 +413,7 @@ describe('Playwright', () => {
       // - C being included
       // - D being included
       // - E being excluded
-      expect(
+      await expect(
         transformPlaywright(
           dedent`
         export default { title: 'foo/bar', component: Button };
@@ -423,261 +425,259 @@ describe('Playwright', () => {
       `,
           filename
         )
-      ).toMatchInlineSnapshot(`
-        if (!require.main) {
-          describe("Example/foo/bar", () => {
-            describe("B", () => {
-              it.skip("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--b",
-                    title: "Example/foo/bar",
-                    name: "B"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--b"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
+      ).resolves.toMatchInlineSnapshot(`
+        describe("Example/Header", () => {
+          describe("B", () => {
+            it.skip("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--b",
+                  title: "Example/Header",
+                  name: "B"
                 };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
                 try {
-                  await testFn();
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--b"
+                  });
                 } catch (err) {
                   if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"B"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
+                    throw err;
                   } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
                     throw err;
                   }
                 }
-              });
-            });
-            describe("C", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--c",
-                    title: "Example/foo/bar",
-                    name: "C"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--c"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
-                  await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"C"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
                 }
-              });
-            });
-            describe("D", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--d",
-                    title: "Example/foo/bar",
-                    name: "D"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
                   }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--d"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
-                  await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"D"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
+                  await jestPlaywright.saveCoverage(page);
                 }
-              });
-            });
-            describe("E", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--e",
-                    title: "Example/foo/bar",
-                    name: "E"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--e"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"B"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
                   await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"E"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
+                } else {
+                  throw err;
                 }
-              });
+              }
             });
           });
-        }
+          describe("C", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--c",
+                  title: "Example/Header",
+                  name: "C"
+                };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
+                try {
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--c"
+                  });
+                } catch (err) {
+                  if (err.toString().includes('Execution context was destroyed')) {
+                    throw err;
+                  } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
+                    throw err;
+                  }
+                }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"C"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
+            });
+          });
+          describe("D", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--d",
+                  title: "Example/Header",
+                  name: "D"
+                };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
+                try {
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--d"
+                  });
+                } catch (err) {
+                  if (err.toString().includes('Execution context was destroyed')) {
+                    throw err;
+                  } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
+                    throw err;
+                  }
+                }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"D"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
+            });
+          });
+          describe("E", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--e",
+                  title: "Example/Header",
+                  name: "E"
+                };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
+                try {
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--e"
+                  });
+                } catch (err) {
+                  if (err.toString().includes('Execution context was destroyed')) {
+                    throw err;
+                  } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
+                    throw err;
+                  }
+                }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"E"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
+            });
+          });
+        });
       `);
     });
-    it('should work with tag negation', () => {
+    it('should work with tag negation', async () => {
       process.env.STORYBOOK_INCLUDE_TAGS = 'play,test';
       process.env.STORYBOOK_PREVIEW_TAGS = '!test';
       // Should result in:
       // - A being included
       // - B being excluded because it has no play nor test tag (removed by negation in preview tags)
       // - C being included because it has test tag (overwritten via story tags)
-      expect(
+      await expect(
         transformPlaywright(
           dedent`
         export default { title: 'foo/bar', component: Button, tags: ['play'] };
@@ -687,138 +687,136 @@ describe('Playwright', () => {
       `,
           filename
         )
-      ).toMatchInlineSnapshot(`
-        if (!require.main) {
-          describe("Example/foo/bar", () => {
-            describe("A", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--a",
-                    title: "Example/foo/bar",
-                    name: "A"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--a"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
+      ).resolves.toMatchInlineSnapshot(`
+        describe("Example/Header", () => {
+          describe("A", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--a",
+                  title: "Example/Header",
+                  name: "A"
                 };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
                 try {
-                  await testFn();
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--a"
+                  });
                 } catch (err) {
                   if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"A"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
+                    throw err;
                   } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
                     throw err;
                   }
                 }
-              });
-            });
-            describe("C", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--c",
-                    title: "Example/foo/bar",
-                    name: "C"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--c"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
-                  await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"C"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
                 }
-              });
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"A"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
             });
           });
-        }
+          describe("C", () => {
+            it("smoke-test", async () => {
+              const testFn = async () => {
+                const context = {
+                  id: "example-header--c",
+                  title: "Example/Header",
+                  name: "C"
+                };
+                if (globalThis.__sbPreVisit) {
+                  await globalThis.__sbPreVisit(page, context);
+                }
+                let result;
+                try {
+                  result = await page.evaluate(({
+                    id,
+                    hasPlayFn
+                  }) => __test(id, hasPlayFn), {
+                    id: "example-header--c"
+                  });
+                } catch (err) {
+                  if (err.toString().includes('Execution context was destroyed')) {
+                    throw err;
+                  } else {
+                    if (globalThis.__sbPostVisit) {
+                      await globalThis.__sbPostVisit(page, {
+                        ...context,
+                        hasFailure: true
+                      });
+                    }
+                    throw err;
+                  }
+                }
+                if (globalThis.__sbPostVisit) {
+                  await globalThis.__sbPostVisit(page, context);
+                }
+                if (globalThis.__sbCollectCoverage) {
+                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                  if (!isCoverageSetupCorrectly) {
+                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+        The code in this story is not instrumented, which means the coverage setup is likely not correct.
+        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                  }
+                  await jestPlaywright.saveCoverage(page);
+                }
+                return result;
+              };
+              try {
+                await testFn();
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"C"}". Retrying...\`);
+                  await jestPlaywright.resetPage();
+                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                  await testFn();
+                } else {
+                  throw err;
+                }
+              }
+            });
+          });
+        });
       `);
     });
-    it('should include "test" tag by default', () => {
+    it('should include "test" tag by default', async () => {
       // Should result in:
       // - A being included
       // - B being excluded
-      expect(
+      await expect(
         transformPlaywright(
           dedent`
         export default { title: 'foo/bar', component: Button };
@@ -827,252 +825,7 @@ describe('Playwright', () => {
       `,
           filename
         )
-      ).toMatchInlineSnapshot(`
-        if (!require.main) {
-          describe("Example/foo/bar", () => {
-            describe("A", () => {
-              it("smoke-test", async () => {
-                const testFn = async () => {
-                  const context = {
-                    id: "example-foo-bar--a",
-                    title: "Example/foo/bar",
-                    name: "A"
-                  };
-                  if (globalThis.__sbPreVisit) {
-                    await globalThis.__sbPreVisit(page, context);
-                  }
-                  let result;
-                  try {
-                    result = await page.evaluate(({
-                      id,
-                      hasPlayFn
-                    }) => __test(id, hasPlayFn), {
-                      id: "example-foo-bar--a"
-                    });
-                  } catch (err) {
-                    if (err.toString().includes('Execution context was destroyed')) {
-                      throw err;
-                    } else {
-                      if (globalThis.__sbPostVisit) {
-                        await globalThis.__sbPostVisit(page, {
-                          ...context,
-                          hasFailure: true
-                        });
-                      }
-                      throw err;
-                    }
-                  }
-                  if (globalThis.__sbPostVisit) {
-                    await globalThis.__sbPostVisit(page, context);
-                  }
-                  if (globalThis.__sbCollectCoverage) {
-                    const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                    if (!isCoverageSetupCorrectly) {
-                      throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-          The code in this story is not instrumented, which means the coverage setup is likely not correct.
-          More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                    }
-                    await jestPlaywright.saveCoverage(page);
-                  }
-                  return result;
-                };
-                try {
-                  await testFn();
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"A"}". Retrying...\`);
-                    await jestPlaywright.resetPage();
-                    await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                    await testFn();
-                  } else {
-                    throw err;
-                  }
-                }
-              });
-            });
-          });
-        }
-      `);
-    });
-    it('should no op when includeTags is passed but not matched', () => {
-      process.env.STORYBOOK_INCLUDE_TAGS = 'play';
-      expect(
-        transformPlaywright(
-          dedent`
-        export default { title: 'foo/bar', component: Button };
-        export const A = () => {};
-        A.play = () => {};
-      `,
-          filename
-        )
-      ).toMatchInlineSnapshot(`describe.skip('Example/foo/bar', () => { it('no-op', () => {}) });`);
-    });
-  });
-
-  it('should generate a play test when the story has a play function', () => {
-    expect(
-      transformPlaywright(
-        dedent`
-        export default { title: 'foo/bar', component: Button };
-        export const A = () => {};
-        A.play = () => {};
-      `,
-        filename
-      )
-    ).toMatchInlineSnapshot(`
-      if (!require.main) {
-        describe("Example/foo/bar", () => {
-          describe("A", () => {
-            it("play-test", async () => {
-              const testFn = async () => {
-                const context = {
-                  id: "example-foo-bar--a",
-                  title: "Example/foo/bar",
-                  name: "A"
-                };
-                if (globalThis.__sbPreVisit) {
-                  await globalThis.__sbPreVisit(page, context);
-                }
-                let result;
-                try {
-                  result = await page.evaluate(({
-                    id,
-                    hasPlayFn
-                  }) => __test(id, hasPlayFn), {
-                    id: "example-foo-bar--a"
-                  });
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    throw err;
-                  } else {
-                    if (globalThis.__sbPostVisit) {
-                      await globalThis.__sbPostVisit(page, {
-                        ...context,
-                        hasFailure: true
-                      });
-                    }
-                    throw err;
-                  }
-                }
-                if (globalThis.__sbPostVisit) {
-                  await globalThis.__sbPostVisit(page, context);
-                }
-                if (globalThis.__sbCollectCoverage) {
-                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                  if (!isCoverageSetupCorrectly) {
-                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-        The code in this story is not instrumented, which means the coverage setup is likely not correct.
-        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                  }
-                  await jestPlaywright.saveCoverage(page);
-                }
-                return result;
-              };
-              try {
-                await testFn();
-              } catch (err) {
-                if (err.toString().includes('Execution context was destroyed')) {
-                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"A"}". Retrying...\`);
-                  await jestPlaywright.resetPage();
-                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                  await testFn();
-                } else {
-                  throw err;
-                }
-              }
-            });
-          });
-        });
-      }
-    `);
-  });
-  it('should generate a smoke test when story does not have a play function', () => {
-    expect(
-      transformPlaywright(
-        dedent`
-        export default { title: 'foo/bar' };
-        export const A = () => {};
-      `,
-        filename
-      )
-    ).toMatchInlineSnapshot(`
-      if (!require.main) {
-        describe("Example/foo/bar", () => {
-          describe("A", () => {
-            it("smoke-test", async () => {
-              const testFn = async () => {
-                const context = {
-                  id: "example-foo-bar--a",
-                  title: "Example/foo/bar",
-                  name: "A"
-                };
-                if (globalThis.__sbPreVisit) {
-                  await globalThis.__sbPreVisit(page, context);
-                }
-                let result;
-                try {
-                  result = await page.evaluate(({
-                    id,
-                    hasPlayFn
-                  }) => __test(id, hasPlayFn), {
-                    id: "example-foo-bar--a"
-                  });
-                } catch (err) {
-                  if (err.toString().includes('Execution context was destroyed')) {
-                    throw err;
-                  } else {
-                    if (globalThis.__sbPostVisit) {
-                      await globalThis.__sbPostVisit(page, {
-                        ...context,
-                        hasFailure: true
-                      });
-                    }
-                    throw err;
-                  }
-                }
-                if (globalThis.__sbPostVisit) {
-                  await globalThis.__sbPostVisit(page, context);
-                }
-                if (globalThis.__sbCollectCoverage) {
-                  const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
-                  if (!isCoverageSetupCorrectly) {
-                    throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
-        The code in this story is not instrumented, which means the coverage setup is likely not correct.
-        More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
-                  }
-                  await jestPlaywright.saveCoverage(page);
-                }
-                return result;
-              };
-              try {
-                await testFn();
-              } catch (err) {
-                if (err.toString().includes('Execution context was destroyed')) {
-                  console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/foo/bar"}/\${"A"}". Retrying...\`);
-                  await jestPlaywright.resetPage();
-                  await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
-                  await testFn();
-                } else {
-                  throw err;
-                }
-              }
-            });
-          });
-        });
-      }
-    `);
-  });
-  it('should generate a smoke test with auto title', () => {
-    expect(
-      transformPlaywright(
-        dedent`
-        export default { component: Button };
-        export const A = () => {};
-      `,
-        filename
-      )
-    ).toMatchInlineSnapshot(`
-      if (!require.main) {
+      ).resolves.toMatchInlineSnapshot(`
         describe("Example/Header", () => {
           describe("A", () => {
             it("smoke-test", async () => {
@@ -1135,7 +888,246 @@ describe('Playwright', () => {
             });
           });
         });
-      }
+      `);
+    });
+    it('should no op when includeTags is passed but not matched', async () => {
+      process.env.STORYBOOK_INCLUDE_TAGS = 'play';
+      await expect(
+        transformPlaywright(
+          dedent`
+        export default { title: 'foo/bar', component: Button };
+        export const A = () => {};
+        A.play = () => {};
+      `,
+          filename
+        )
+      ).resolves.toMatchInlineSnapshot(
+        `describe.skip('Example/Header', () => { it('no-op', () => {}) });`
+      );
+    });
+  });
+
+  it('should generate a play test when the story has a play function', async () => {
+    await expect(
+      transformPlaywright(
+        dedent`
+        export default { title: 'foo/bar', component: Button };
+        export const A = () => {};
+        A.play = () => {};
+      `,
+        filename
+      )
+    ).resolves.toMatchInlineSnapshot(`
+      describe("Example/Header", () => {
+        describe("A", () => {
+          it("play-test", async () => {
+            const testFn = async () => {
+              const context = {
+                id: "example-header--a",
+                title: "Example/Header",
+                name: "A"
+              };
+              if (globalThis.__sbPreVisit) {
+                await globalThis.__sbPreVisit(page, context);
+              }
+              let result;
+              try {
+                result = await page.evaluate(({
+                  id,
+                  hasPlayFn
+                }) => __test(id, hasPlayFn), {
+                  id: "example-header--a"
+                });
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  throw err;
+                } else {
+                  if (globalThis.__sbPostVisit) {
+                    await globalThis.__sbPostVisit(page, {
+                      ...context,
+                      hasFailure: true
+                    });
+                  }
+                  throw err;
+                }
+              }
+              if (globalThis.__sbPostVisit) {
+                await globalThis.__sbPostVisit(page, context);
+              }
+              if (globalThis.__sbCollectCoverage) {
+                const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                if (!isCoverageSetupCorrectly) {
+                  throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+      The code in this story is not instrumented, which means the coverage setup is likely not correct.
+      More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                }
+                await jestPlaywright.saveCoverage(page);
+              }
+              return result;
+            };
+            try {
+              await testFn();
+            } catch (err) {
+              if (err.toString().includes('Execution context was destroyed')) {
+                console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"A"}". Retrying...\`);
+                await jestPlaywright.resetPage();
+                await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                await testFn();
+              } else {
+                throw err;
+              }
+            }
+          });
+        });
+      });
+    `);
+  });
+  it('should generate a smoke test when story does not have a play function', async () => {
+    await expect(
+      transformPlaywright(
+        dedent`
+        export default { title: 'foo/bar' };
+        export const A = () => {};
+      `,
+        filename
+      )
+    ).resolves.toMatchInlineSnapshot(`
+      describe("Example/Header", () => {
+        describe("A", () => {
+          it("smoke-test", async () => {
+            const testFn = async () => {
+              const context = {
+                id: "example-header--a",
+                title: "Example/Header",
+                name: "A"
+              };
+              if (globalThis.__sbPreVisit) {
+                await globalThis.__sbPreVisit(page, context);
+              }
+              let result;
+              try {
+                result = await page.evaluate(({
+                  id,
+                  hasPlayFn
+                }) => __test(id, hasPlayFn), {
+                  id: "example-header--a"
+                });
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  throw err;
+                } else {
+                  if (globalThis.__sbPostVisit) {
+                    await globalThis.__sbPostVisit(page, {
+                      ...context,
+                      hasFailure: true
+                    });
+                  }
+                  throw err;
+                }
+              }
+              if (globalThis.__sbPostVisit) {
+                await globalThis.__sbPostVisit(page, context);
+              }
+              if (globalThis.__sbCollectCoverage) {
+                const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                if (!isCoverageSetupCorrectly) {
+                  throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+      The code in this story is not instrumented, which means the coverage setup is likely not correct.
+      More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                }
+                await jestPlaywright.saveCoverage(page);
+              }
+              return result;
+            };
+            try {
+              await testFn();
+            } catch (err) {
+              if (err.toString().includes('Execution context was destroyed')) {
+                console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"A"}". Retrying...\`);
+                await jestPlaywright.resetPage();
+                await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                await testFn();
+              } else {
+                throw err;
+              }
+            }
+          });
+        });
+      });
+    `);
+  });
+  it('should generate a smoke test with auto title', async () => {
+    await expect(
+      transformPlaywright(
+        dedent`
+        export default { component: Button };
+        export const A = () => {};
+      `,
+        filename
+      )
+    ).resolves.toMatchInlineSnapshot(`
+      describe("Example/Header", () => {
+        describe("A", () => {
+          it("smoke-test", async () => {
+            const testFn = async () => {
+              const context = {
+                id: "example-header--a",
+                title: "Example/Header",
+                name: "A"
+              };
+              if (globalThis.__sbPreVisit) {
+                await globalThis.__sbPreVisit(page, context);
+              }
+              let result;
+              try {
+                result = await page.evaluate(({
+                  id,
+                  hasPlayFn
+                }) => __test(id, hasPlayFn), {
+                  id: "example-header--a"
+                });
+              } catch (err) {
+                if (err.toString().includes('Execution context was destroyed')) {
+                  throw err;
+                } else {
+                  if (globalThis.__sbPostVisit) {
+                    await globalThis.__sbPostVisit(page, {
+                      ...context,
+                      hasFailure: true
+                    });
+                  }
+                  throw err;
+                }
+              }
+              if (globalThis.__sbPostVisit) {
+                await globalThis.__sbPostVisit(page, context);
+              }
+              if (globalThis.__sbCollectCoverage) {
+                const isCoverageSetupCorrectly = await page.evaluate(() => '__coverage__' in window);
+                if (!isCoverageSetupCorrectly) {
+                  throw new Error(\`[Test runner] An error occurred when evaluating code coverage:
+      The code in this story is not instrumented, which means the coverage setup is likely not correct.
+      More info: https://github.com/storybookjs/test-runner#setting-up-code-coverage\`);
+                }
+                await jestPlaywright.saveCoverage(page);
+              }
+              return result;
+            };
+            try {
+              await testFn();
+            } catch (err) {
+              if (err.toString().includes('Execution context was destroyed')) {
+                console.log(\`An error occurred in the following story, most likely because of a navigation: "\${"Example/Header"}/\${"A"}". Retrying...\`);
+                await jestPlaywright.resetPage();
+                await globalThis.__sbSetupPage(globalThis.page, globalThis.context);
+                await testFn();
+              } else {
+                throw err;
+              }
+            }
+          });
+        });
+      });
     `);
   });
 });
