@@ -1,9 +1,12 @@
 import { relative } from 'path';
-import template from '@babel/template';
+import babelTemplate from '@babel/template';
+
+// Handle both ESM and CJS patterns
+const template = (babelTemplate as any).default ?? babelTemplate;
 import { userOrAutoTitle } from 'storybook/internal/preview-api';
 import dedent from 'ts-dedent';
 
-import { getStorybookMetadata } from '../util';
+import { getStorybookMetadata } from '../util/getStorybookMetadata';
 import { transformCsf } from '../csf/transformCsf';
 import type { TestPrefixer } from '../csf/transformCsf';
 
@@ -79,24 +82,24 @@ export const testPrefixer: TestPrefixer = (context) => {
   )({ ...context });
 };
 
-const makeTitleFactory = (filename: string) => {
-  const { workingDir, normalizedStoriesEntries } = getStorybookMetadata();
+const makeTitleFactory = async (filename: string) => {
+  const { workingDir, normalizedStoriesEntries } = await getStorybookMetadata();
   const filePath = `./${relative(workingDir, filename)}`;
 
   return (userTitle: string) =>
     userOrAutoTitle(filePath, normalizedStoriesEntries, userTitle) as string;
 };
 
-export const transformPlaywright = (src: string, filename: string) => {
+export const transformPlaywright = async (src: string, filename: string) => {
   const tags = process.env.STORYBOOK_PREVIEW_TAGS?.split(',') ?? [];
   const transformOptions = {
     testPrefixer,
     insertTestIfEmpty: true,
     clearBody: true,
-    makeTitle: makeTitleFactory(filename),
+    makeTitle: await makeTitleFactory(filename),
     previewAnnotations: { tags },
   };
 
-  const result = transformCsf(src, transformOptions);
+  const result = await transformCsf(src, transformOptions);
   return result;
 };
