@@ -4,7 +4,8 @@ import { pkgUp } from 'pkg-up';
 import { PrepareContext } from './playwright/hooks';
 import { getTestRunnerConfig } from './util/getTestRunnerConfig';
 import { readFile } from 'node:fs/promises';
-import path from 'node:path';
+import path, { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
 
 /**
  * This is a default prepare function which can be overridden by the user.
@@ -36,7 +37,10 @@ export const setupPage = async (page: Page, browserContext: BrowserContext) => {
   const failOnConsole = process.env.TEST_CHECK_CONSOLE;
 
   const viewMode = process.env.VIEW_MODE ?? 'story';
-  const renderedEvent = viewMode === 'docs' ? 'globalThis.__STORYBOOK_MODULE_CORE_EVENTS__.DOCS_RENDERED' : 'globalThis.__STORYBOOK_MODULE_CORE_EVENTS__.STORY_FINISHED ?? globalThis.__STORYBOOK_MODULE_CORE_EVENTS__.STORY_RENDERED';
+  const renderedEvent =
+    viewMode === 'docs'
+      ? 'globalThis.__STORYBOOK_MODULE_CORE_EVENTS__.DOCS_RENDERED'
+      : 'globalThis.__STORYBOOK_MODULE_CORE_EVENTS__.STORY_FINISHED ?? globalThis.__STORYBOOK_MODULE_CORE_EVENTS__.STORY_RENDERED';
   const { packageJson } = (await readPackageUp()) as NormalizedReadResult;
   const { version: testRunnerVersion } = packageJson;
 
@@ -51,7 +55,7 @@ export const setupPage = async (page: Page, browserContext: BrowserContext) => {
     );
   }
 
-  const testRunnerConfig = getTestRunnerConfig() || {};
+  const testRunnerConfig = (await getTestRunnerConfig()) || {};
   if (testRunnerConfig?.prepare) {
     await testRunnerConfig.prepare({ page, browserContext, testRunnerConfig });
   } else {
@@ -69,16 +73,18 @@ export const setupPage = async (page: Page, browserContext: BrowserContext) => {
     return message;
   });
 
+  const cwd = typeof __dirname === 'string' ? __dirname : dirname(fileURLToPath(import.meta.url));
+
   const finalStorybookUrl = referenceURL ?? targetURL ?? '';
-  const testRunnerPackageLocation = await pkgUp({ cwd: __dirname });
+  const testRunnerPackageLocation = await pkgUp({ cwd });
   if (!testRunnerPackageLocation) throw new Error('Could not find test-runner package location');
   const scriptLocation = path.join(
     path.dirname(testRunnerPackageLocation),
     'dist',
-    'setup-page-script.mjs'
+    'setup-page-script.js'
   );
 
-  // read the content of setup-page-script.mjs and replace the placeholders with the actual values
+  // read the content of setup-page-script.js and replace the placeholders with the actual values
   const content = (await readFile(scriptLocation, 'utf-8'))
     .replaceAll('{{storybookUrl}}', finalStorybookUrl)
     .replaceAll('{{failOnConsole}}', failOnConsole ?? 'false')
