@@ -1,32 +1,39 @@
 import { getStorybookMain, resetStorybookMainCache, storybookMainConfig } from './getStorybookMain';
-import * as coreCommon from 'storybook/internal/common';
+import { describe, it, expect, beforeEach, vi } from 'vitest';
+import { serverRequire } from 'storybook/internal/common';
 
-jest.mock('storybook/internal/common');
+vi.mock('storybook/internal/common', () => ({
+  serverRequire: vi.fn(),
+}));
 
 describe('getStorybookMain', () => {
   beforeEach(() => {
     resetStorybookMainCache();
+    vi.clearAllMocks();
   });
 
-  it('should throw an error if no configuration is found', () => {
-    expect(() => getStorybookMain('.storybook')).toThrowErrorMatchingSnapshot();
+  it('should throw an error if no configuration is found', async () => {
+    vi.mocked(serverRequire).mockRejectedValueOnce(new Error('Module not found'));
+    await expect(getStorybookMain('.storybook')).rejects.toThrowErrorMatchingSnapshot();
   });
 
   describe('no stories', () => {
-    it('should throw an error if no stories are defined', () => {
-      jest.spyOn(coreCommon, 'serverRequire').mockImplementation(() => ({}));
+    it('should throw an error if no stories are defined', async () => {
+      vi.mocked(serverRequire).mockResolvedValueOnce({});
 
-      expect(() => getStorybookMain('.storybook')).toThrowErrorMatchingSnapshot();
+      await expect(getStorybookMain('.storybook')).rejects.toThrowErrorMatchingSnapshot();
     });
 
-    it('should throw an error if stories list is empty', () => {
-      jest.spyOn(coreCommon, 'serverRequire').mockImplementation(() => ({ stories: [] }));
+    it('should throw an error if stories list is empty', async () => {
+      vi.mocked(serverRequire).mockResolvedValueOnce({
+        stories: [],
+      });
 
-      expect(() => getStorybookMain('.storybook')).toThrowErrorMatchingSnapshot();
+      await expect(getStorybookMain('.storybook')).rejects.toThrowErrorMatchingSnapshot();
     });
   });
 
-  it('should return mainjs', () => {
+  it('should return mainjs', async () => {
     const mockedMain = {
       stories: [
         {
@@ -36,13 +43,13 @@ describe('getStorybookMain', () => {
       ],
     };
 
-    jest.spyOn(coreCommon, 'serverRequire').mockImplementation(() => mockedMain);
+    vi.mocked(serverRequire).mockResolvedValueOnce(mockedMain);
 
-    const res = getStorybookMain('.storybook');
+    const res = await getStorybookMain('.storybook');
     expect(res).toMatchObject(mockedMain);
   });
 
-  it('should return the configDir value if it exists', () => {
+  it('should return the configDir value if it exists', async () => {
     const mockedMain = {
       stories: [
         {
@@ -51,9 +58,9 @@ describe('getStorybookMain', () => {
         },
       ],
     };
-    storybookMainConfig.set('configDir', mockedMain);
+    storybookMainConfig.set('.storybook', mockedMain);
 
-    const res = getStorybookMain('.storybook');
+    const res = await getStorybookMain('.storybook');
     expect(res).toMatchObject(mockedMain);
   });
 });
