@@ -4,11 +4,11 @@ import { StoriesEntry } from 'storybook/internal/types';
 
 import { getStorybookMain } from './getStorybookMain';
 
-export const getStorybookMetadata = () => {
+export const getStorybookMetadata = async () => {
   const workingDir = getProjectRoot();
   const configDir = process.env.STORYBOOK_CONFIG_DIR ?? '.storybook';
 
-  const main = getStorybookMain(configDir);
+  const main = await getStorybookMain(configDir);
   const normalizedStoriesEntries = normalizeStories(main.stories as StoriesEntry[], {
     configDir,
     workingDir,
@@ -22,10 +22,8 @@ export const getStorybookMetadata = () => {
     .map((dir) => join(workingDir, dir))
     .join(';');
 
-  // @ts-expect-error -- this is added in storybook/internal/common@6.5, which we don't depend on
   const lazyCompilation = !!main.core?.builder?.options?.lazyCompilation;
 
-  // @ts-expect-error -- need to update to latest sb version
   const { disableTelemetry, enableCrashReports } = main.core || {};
 
   return {
